Refactored adapter out of fragment and reimplemented interaction listeners in a more standard android fashion
