+++ box-browse-sdk/src/main/java/com/box/androidsdk/browse/fragments/BoxBrowseFragment.java
@@ -6,15 +6,12 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.res.Resources;
 import android.graphics.Canvas;
-import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentTransaction;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.ActionBar;
@@ -26,51 +23,37 @@ import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CheckBox;
-import android.widget.ImageButton;
-import android.widget.ImageView;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import com.box.androidsdk.browse.R;
-import com.box.androidsdk.browse.activities.BoxBrowseActivity;
+import com.box.androidsdk.browse.adapters.BoxItemAdapter;
 import com.box.androidsdk.browse.filters.BoxItemFilter;
 import com.box.androidsdk.browse.service.BoxBrowseController;
 import com.box.androidsdk.browse.service.BoxResponseIntent;
 import com.box.androidsdk.browse.service.BrowseController;
 import com.box.androidsdk.browse.service.CompletionListener;
 import com.box.androidsdk.browse.uidata.BoxListItem;
-import com.box.androidsdk.browse.uidata.ThumbnailManager;
 import com.box.androidsdk.content.BoxApiFile;
 import com.box.androidsdk.content.BoxApiFolder;
 import com.box.androidsdk.content.BoxApiSearch;
-import com.box.androidsdk.content.BoxException;
-import com.box.androidsdk.content.models.BoxFile;
-import com.box.androidsdk.content.models.BoxFolder;
 import com.box.androidsdk.content.models.BoxItem;
 import com.box.androidsdk.content.models.BoxIteratorItems;
 import com.box.androidsdk.content.models.BoxSession;
 import com.box.androidsdk.content.requests.BoxRequestsFile;
-import com.box.androidsdk.content.utils.BoxLogUtils;
 import com.box.androidsdk.content.utils.SdkUtils;
 
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.Serializable;
 import java.lang.ref.WeakReference;
-import java.net.HttpURLConnection;
-import java.text.DateFormat;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 
 /**
  * Base fragment for displaying box items. This class provides the internals needed to make requests
  * and update the ui on a response.
  */
-public abstract class BoxBrowseFragment extends Fragment implements SwipeRefreshLayout.OnRefreshListener {
+public abstract class BoxBrowseFragment extends Fragment implements SwipeRefreshLayout.OnRefreshListener,
+        BoxItemAdapter.OnInteractionListener {
     public static final String TAG = BoxBrowseFragment.class.getName();
 
     protected static final int DEFAULT_LIMIT = 200;
@@ -85,21 +68,19 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
     protected static final String EXTRA_MULTI_SELECT_HANDLER = "com.box.androidsdk.browse.MULTI_SELECT_HANDLER";
     protected static final String EXTRA_TITLE = "com.box.androidsdk.browse.TITLE";
     protected static final String EXTRA_COLLECTION = "com.box.androidsdk.browse.COLLECTION";
-    protected static final String ACTION_FUTURE_TASK = "com.box.androidsdk.browse.FUTURE_TASK";
+    public static final String ACTION_FUTURE_TASK = "com.box.androidsdk.browse.FUTURE_TASK";
 
     protected BoxSession mSession;
     protected BoxIteratorItems mBoxIteratorItems;
 
-    protected OnFragmentInteractionListener mListener;
-    protected OnFragmentInteractionListener mSecondaryActionListener;
+    protected OnItemClickListener mListener;
+    protected OnSecondaryActionListener mSecondaryActionListener;
     protected MultiSelectHandler mMultiSelectHandler;
 
     protected BoxItemAdapter mAdapter;
     protected RecyclerView mItemsView;
-    protected ThumbnailManager mThumbnailManager;
     protected SwipeRefreshLayout mSwipeRefresh;
     protected ProgressBar mProgress;
-    protected LocalBroadcastManager mLocalBroadcastManager;
     protected int mLimit = DEFAULT_LIMIT;
 
     private String mTitle;
@@ -130,20 +111,10 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
     };
     private View mRootView;
     private BoxItemFilter mBoxItemFilter;
+    private LocalBroadcastManager mLocalBroadcastmanager;
 
     public BoxBrowseFragment() {
         // Required empty public constructor
-    }
-
-    public static void setAccentColor(Resources res, ProgressBar progressBar) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            int accentColor = res.getColor(R.color.box_accent);
-            Drawable drawable = progressBar.getIndeterminateDrawable();
-            if (drawable != null) {
-                drawable.setColorFilter(accentColor, PorterDuff.Mode.SRC_IN);
-                drawable.invalidateSelf();
-            }
-        }
     }
 
     @Override
@@ -155,22 +126,18 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
                 throw new IllegalArgumentException("A valid session or user id must be provided");
             }
             mSession = new BoxSession(getActivity(), userId);
-            mThumbnailManager = initializeThumbnailManager();
             mLimit = getArguments().getInt(ARG_LIMIT);
             mBoxItemFilter = (BoxItemFilter) getArguments().getSerializable(ARG_BOX_ITEM_FILTER);
         }
         if (savedInstanceState != null) {
             setListItems((BoxIteratorItems) savedInstanceState.getSerializable(EXTRA_COLLECTION));
             if (savedInstanceState.containsKey(EXTRA_SECONDARY_ACTION_LISTENER)) {
-                mSecondaryActionListener = (OnFragmentInteractionListener) savedInstanceState.getSerializable(EXTRA_SECONDARY_ACTION_LISTENER);
+                mSecondaryActionListener = (OnSecondaryActionListener) savedInstanceState.getSerializable(EXTRA_SECONDARY_ACTION_LISTENER);
             }
             if (savedInstanceState.containsKey(EXTRA_MULTI_SELECT_HANDLER)) {
                 mMultiSelectHandler = (MultiSelectHandler) savedInstanceState.getSerializable(EXTRA_MULTI_SELECT_HANDLER);
             }
         }
-
-        // Initialize controller and listeners
-        mLocalBroadcastManager = LocalBroadcastManager.getInstance(getActivity());
 
         // TODO: Do we really need this?
         setRetainInstance(true);
@@ -179,13 +146,13 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
     @Override
     public void onResume() {
         getActivity().registerReceiver(mConnectivityReceiver, new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION));
-        mLocalBroadcastManager.registerReceiver(mBroadcastReceiver, getIntentFilter());
+        getLocalBroadcastManager().registerReceiver(mBroadcastReceiver, getIntentFilter());
         super.onResume();
     }
 
     @Override
     public void onPause() {
-        mLocalBroadcastManager.unregisterReceiver(mBroadcastReceiver);
+        getLocalBroadcastManager().unregisterReceiver(mBroadcastReceiver);
         getActivity().unregisterReceiver(mConnectivityReceiver);
         super.onPause();
     }
@@ -208,15 +175,6 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         super.onActivityCreated(savedInstanceState);
         if (savedInstanceState != null) {
             setToolbar(savedInstanceState.getString(EXTRA_TITLE));
-        }
-    }
-
-    private ThumbnailManager initializeThumbnailManager() {
-        try {
-            return new ThumbnailManager(getController().getThumbnailCacheDir());
-        } catch (FileNotFoundException e) {
-            BoxLogUtils.e(TAG, e);
-            return null;
         }
     }
 
@@ -252,7 +210,7 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         mItemsView.addItemDecoration(new BoxItemDividerDecoration(getResources()));
         mItemsView.setLayoutManager(new LinearLayoutManager(getActivity()));
         mProgress = (ProgressBar) mRootView.findViewById(R.id.box_browsesdk_progress_bar);
-        mAdapter = createBoxItemAdapter();
+        mAdapter = createAdapter();
         mItemsView.setAdapter(mAdapter);
         if (getMultiSelectHandler() != null) {
             getMultiSelectHandler().setItemAdapter(mAdapter);
@@ -267,23 +225,23 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         return mRootView;
     }
 
-    protected BoxItemAdapter createBoxItemAdapter(){
-        return new BoxItemAdapter();
-    }
-
     protected abstract void loadItems();
 
     protected void setListItems(final BoxIteratorItems items) {
         mBoxIteratorItems = items;
     }
 
+    protected BoxItemAdapter createAdapter() {
+        return new BoxItemAdapter(getActivity(), mBoxItemFilter, getController(), this);
+    }
+
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
         try {
-            mListener = (OnFragmentInteractionListener) context;
+            mListener = (OnItemClickListener) context;
         } catch (ClassCastException e) {
-            throw new ClassCastException(context.toString() + " must implement OnFragmentInteractionListener");
+            throw new ClassCastException(context.toString() + " must implement OnItemClickListener");
         }
     }
 
@@ -302,7 +260,7 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
     public BrowseController getController() {
         if (mController == null) {
             mController = new BoxBrowseController(mSession, new BoxApiFile(mSession), new BoxApiFolder(mSession), new BoxApiSearch(mSession))
-                    .setCompletedListener(new CompletionListener(mLocalBroadcastManager));
+                    .setCompletedListener(new CompletionListener(getLocalBroadcastManager()));
         }
         return mController;
     }
@@ -311,17 +269,22 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         mController = controller;
     }
 
-    protected OnFragmentInteractionListener getSecondaryActionListener() {
+    public OnSecondaryActionListener getOnSecondaryActionListener() {
         return mSecondaryActionListener;
+    }
+
+    @Override
+    public OnItemClickListener getOnItemClickListener() {
+        return (OnItemClickListener) getActivity();
     }
 
     /**
      * Optionally set a secondary action to use on this fragment.
      *
      * @param listener listener to be called when a secondary action is clicked on an item. Must be serializable.
-     * @param <T>      Serializable OnFragmentInteractionListener.
+     * @param <T>      Serializable OnSecondaryActionListener.
      */
-    public <T extends OnFragmentInteractionListener & Serializable> void setSecondaryActionListener(T listener) {
+    public <T extends OnSecondaryActionListener & Serializable> void setSecondaryActionListener(T listener) {
         mSecondaryActionListener = listener;
         if (mAdapter != null) {
             mAdapter.notifyItemRangeChanged(0, mAdapter.getItemCount());
@@ -331,6 +294,7 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
     /**
      * @return the MultiSelectHandler set on this fragment. This is used in order to handle batch operations.
      */
+    @Override
     public MultiSelectHandler getMultiSelectHandler() {
         return mMultiSelectHandler;
     }
@@ -339,13 +303,20 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
      * Optionally set a multi select handler.
      *
      * @param handler handler to be called when multiple items are selected.
-     * @param <T>     Serializable OnFragmentInteractionListener.
+     * @param <T>     Serializable OnSecondaryActionListener.
      */
     public <T extends MultiSelectHandler & Serializable> void setMultiSelectHandler(T handler) {
         mMultiSelectHandler = handler;
         if (mAdapter != null) {
             mMultiSelectHandler.setItemAdapter(mAdapter);
         }
+    }
+
+    protected LocalBroadcastManager getLocalBroadcastManager() {
+        if (mLocalBroadcastmanager == null) {
+            mLocalBroadcastmanager = LocalBroadcastManager.getInstance(getActivity());
+        }
+        return mLocalBroadcastmanager;
     }
 
     /**
@@ -415,143 +386,12 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         return filter;
     }
 
-    /**
-     * Called when a {@link BoxListItem} is bound to a ViewHolder. Customizations of UI elements
-     * should be done by overriding this method. If extending from a {@link BoxBrowseActivity}
-     * a custom BoxBrowseFolder fragment can be returned in
-     * {@link BoxBrowseActivity#createBrowseFolderFragment(BoxItem, BoxSession)}
-     *
-     * @param holder the BoxItemHolder
-     */
-    protected void onBindBoxItemViewHolder(BoxItemViewHolder holder) {
-        if (holder.getItem() == null || holder.getItem().getBoxItem() == null) {
-            return;
-        }
-
-        final BoxItem item = holder.getItem().getBoxItem();
-        holder.getNameView().setText(item.getName());
-        String description = "";
-        if (item != null) {
-            String modifiedAt = item.getModifiedAt() != null ?
-                    DateFormat.getDateInstance(DateFormat.MEDIUM).format(item.getModifiedAt()).toUpperCase() :
-                    "";
-            String size = item.getSize() != null ?
-                    localFileSizeToDisplay(item.getSize()) :
-                    "";
-            description = String.format(Locale.ENGLISH, "%s  • %s", modifiedAt, size);
-            mThumbnailManager.setThumbnailIntoView(holder.getThumbView(), item);
-        }
-        holder.getMetaDescription().setText(description);
-        holder.getProgressBar().setVisibility(View.GONE);
-        holder.getMetaDescription().setVisibility(View.VISIBLE);
-        holder.getThumbView().setVisibility(View.VISIBLE);
-        if (!holder.getItem().getIsEnabled()) {
-            holder.getView().setEnabled(false);
-            holder.getNameView().setTextColor(getResources().getColor(R.color.box_browsesdk_hint));
-            holder.getMetaDescription().setTextColor(getResources().getColor(R.color.box_browsesdk_disabled_hint));
-            holder.getThumbView().setAlpha(0.26f);
-        } else {
-            holder.getView().setEnabled(true);
-            holder.getNameView().setTextColor(getResources().getColor(R.color.box_browsesdk_primary_text));
-            holder.getMetaDescription().setTextColor(getResources().getColor(R.color.box_browsesdk_hint));
-            holder.getThumbView().setAlpha(1f);
-        }
-        if (getSecondaryActionListener() != null) {
-            holder.getSecondaryAction().setVisibility(View.VISIBLE);
-        } else {
-            holder.getSecondaryAction().setVisibility(View.GONE);
-        }
-
-        if (getMultiSelectHandler() != null && getMultiSelectHandler().isEnabled()) {
-            holder.getSecondaryAction().setVisibility(View.GONE);
-            holder.getCheckBox().setVisibility(View.VISIBLE);
-            holder.getCheckBox().setEnabled(getMultiSelectHandler().isSelectable(item));
-            holder.getCheckBox().setChecked(getMultiSelectHandler().isItemSelected(item));
-        } else {
-            holder.getCheckBox().setVisibility(View.GONE);
-        }
-
+    public interface OnSecondaryActionListener {
+        boolean onSecondaryAction(BoxItem item);
     }
 
-    /**
-     * Defines the conditions for when a BoxItem should be shown as enabled
-     *
-     * @param item the BoxItem that should be enabled or not
-     * @return whether or not the BoxItem should be enabled
-     */
-    public boolean isItemEnabled(BoxItem item) {
-        return true;
-    }
-
-    /**
-     * Defines the conditions for when a BoxItem should be shown in the adapter
-     *
-     * @param item the BoxItem that should be visible or not
-     * @return whether or not the BoxItem should be visible
-     */
-    public boolean isItemVisible(BoxItem item) {
-        return true;
-    }
-
-    /**
-     * Download the thumbnail for a given file.
-     *
-     * @param fileId file id to download thumbnail for.
-     * @return A FutureTask that is tasked with fetching information on the given folder.
-     */
-    protected BoxRequestsFile.DownloadThumbnail getDownloadThumbnailTask(final String fileId, final File downloadLocation) {
-        return getController().getThumbnailRequest(fileId, downloadLocation, getResources());
-    }
-
-    /**
-     * Java version of routine to turn a long into a short user readable string.
-     * <p/>
-     * This routine is used if the JNI native C version is not available.
-     *
-     * @param numSize the number of bytes in the file.
-     * @return String Short human readable String e.g. 2.5 MB
-     */
-    private String localFileSizeToDisplay(final double numSize) {
-        final int constKB = 1024;
-        final int constMB = constKB * constKB;
-        final int constGB = constMB * constKB;
-        final double floatKB = 1024.0f;
-        final double floatMB = floatKB * floatKB;
-        final double floatGB = floatMB * floatKB;
-        final String BYTES = "B";
-        String textSize = "0 bytes";
-        String strSize = Double.toString(numSize);
-        double size;
-
-        if (numSize < constKB) {
-            textSize = strSize + " " + BYTES;
-        } else if ((numSize >= constKB) && (numSize < constMB)) {
-            size = numSize / floatKB;
-            textSize = String.format(Locale.ENGLISH, "%4.1f KB", size);
-        } else if ((numSize >= constMB) && (numSize < constGB)) {
-            size = numSize / floatMB;
-            textSize = String.format(Locale.ENGLISH, "%4.1f MB", size);
-        } else if (numSize >= constGB) {
-            size = numSize / floatGB;
-            textSize = String.format(Locale.ENGLISH, "%4.1f GB", size);
-        }
-        return textSize;
-    }
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an item being tapped to be communicated to the activity
-     */
-    public interface OnFragmentInteractionListener {
-
-        /**
-         * Called whenever an item in the RecyclerView is clicked and allows the activity to intercept
-         * and override the behavior
-         *
-         * @param item the item that was clicked
-         * @return whether the click event should continue to be handled by the fragment
-         */
-        boolean handleOnItemClick(BoxItem item);
+    public interface OnItemClickListener {
+        void onItemClick(BoxItem item);
     }
 
     public static abstract class MultiSelectHandler {
@@ -603,7 +443,7 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
          */
         public abstract void handleItemSelected(BoxItem boxItem, boolean wasSelected, MultiSelectHandler handler);
 
-        void toggle(BoxItem boxItem) {
+        public void toggle(BoxItem boxItem) {
             if (boxItem == null || !isSelectable(boxItem)) {
                 return;
             }
@@ -624,22 +464,21 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
         /**
          * Select all known box items and update ui.
          */
-        public void selectAll(){
+        public void selectAll() {
             int totalItemCount = mItemAdapter.get().getItemCount();
             if (mSelectedItems.size() < totalItemCount) {
                 int originalSize = mSelectedItems.size();
-                for (BoxListItem item : mItemAdapter.get().mListItems){
+                for (BoxListItem item : mItemAdapter.get().getItems()) {
                     BoxItem boxItem = item.getBoxItem();
-                    if (boxItem != null && isSelectable(boxItem) && !isItemSelected(boxItem) ){
+                    if (boxItem != null && isSelectable(boxItem) && !isItemSelected(boxItem)) {
                         mSelectedItems.add(boxItem);
                         handleItemSelected(boxItem, true, this);
                     }
                 }
-                if (originalSize != mSelectedItems.size()){
+                if (originalSize != mSelectedItems.size()) {
                     mItemAdapter.get().notifyItemRangeChanged(0, mItemAdapter.get().getItemCount());
                 }
             }
-
         }
 
         /**
@@ -650,10 +489,9 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
                 mSelectedItems.clear();
                 mItemAdapter.get().notifyItemRangeChanged(0, mItemAdapter.get().getItemCount());
             }
+         }
 
-        }
-
-        /**
+         /**
          * Return true if multi selecting, false otherwise.
          */
         public boolean isEnabled() {
@@ -704,300 +542,6 @@ public abstract class BoxBrowseFragment extends Fragment implements SwipeRefresh
             }
         }
     }
-
-    protected class BoxItemViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener, View.OnLongClickListener {
-        BoxListItem mItem;
-
-        View mView;
-        ImageView mThumbView;
-        TextView mNameView;
-        TextView mMetaDescription;
-        ProgressBar mProgressBar;
-        ImageButton mSecondaryAction;
-        BoxItemClickListener mSecondaryClickListener;
-        AppCompatCheckBox mItemCheckBox;
-
-
-        public BoxItemViewHolder(View itemView) {
-            super(itemView);
-            itemView.setOnClickListener(this);
-            if (getMultiSelectHandler() != null) {
-                itemView.setOnLongClickListener(this);
-            }
-            mView = itemView;
-            mThumbView = (ImageView) itemView.findViewById(R.id.box_browsesdk_thumb_image);
-            mNameView = (TextView) itemView.findViewById(R.id.box_browsesdk_name_text);
-            mMetaDescription = (TextView) itemView.findViewById(R.id.metaline_description);
-            mProgressBar = (ProgressBar) itemView.findViewById((R.id.spinner));
-            mSecondaryAction = (ImageButton) itemView.findViewById(R.id.secondaryAction);
-            mItemCheckBox = (AppCompatCheckBox) itemView.findViewById(R.id.boxItemCheckBox);
-            mSecondaryClickListener = new BoxItemClickListener();
-            mSecondaryAction.setOnClickListener(mSecondaryClickListener);
-            setAccentColor(getResources(), mProgressBar);
-        }
-
-        public void bindItem(BoxListItem item) {
-            mItem = item;
-            mSecondaryClickListener.setListItem(item);
-            onBindBoxItemViewHolder(this);
-        }
-
-        public void setError(BoxListItem item) {
-            mItem = item;
-            mItem.setState(BoxListItem.State.ERROR);
-            FragmentActivity activity = getActivity();
-            if (activity == null) {
-                return;
-            }
-
-            mThumbView.setImageResource(R.drawable.ic_box_browsesdk_refresh_grey_36dp);
-            mThumbView.setVisibility(View.VISIBLE);
-            mProgressBar.setVisibility(View.GONE);
-            mMetaDescription.setVisibility(View.VISIBLE);
-            mNameView.setText(activity.getResources().getString(R.string.box_browsesdk_error_retrieving_items));
-            mMetaDescription.setText(activity.getResources().getString(R.string.box_browsesdk_tap_to_retry));
-        }
-
-        public void setLoading() {
-            FragmentActivity activity = getActivity();
-            if (activity == null) {
-                return;
-            }
-            mThumbView.setVisibility(View.GONE);
-            mProgressBar.setVisibility(View.VISIBLE);
-            mMetaDescription.setVisibility(View.GONE);
-            mNameView.setText(activity.getResources().getString(R.string.boxsdk_Please_wait));
-        }
-
-        public CheckBox getCheckBox() {
-            return mItemCheckBox;
-        }
-
-        public ImageButton getSecondaryAction() {
-            return mSecondaryAction;
-        }
-
-
-        public BoxListItem getItem() {
-            return mItem;
-        }
-
-        public ProgressBar getProgressBar() {
-            return mProgressBar;
-        }
-
-        public TextView getMetaDescription() {
-            return mMetaDescription;
-        }
-
-        public TextView getNameView() {
-            return mNameView;
-        }
-
-        public ImageView getThumbView() {
-            return mThumbView;
-        }
-
-        public View getView() {
-            return mView;
-        }
-
-
-        @Override
-        public boolean onLongClick(View v) {
-            if (getMultiSelectHandler() != null) {
-                getMultiSelectHandler().setEnabled(!getMultiSelectHandler().isEnabled());
-                getMultiSelectHandler().toggle(mItem.getBoxItem());
-                return true;
-            }
-            return false;
-        }
-
-        @Override
-        public void onClick(View v) {
-            if (getMultiSelectHandler() != null && getMultiSelectHandler().isEnabled()) {
-                getMultiSelectHandler().toggle(mItem.getBoxItem());
-                onBindBoxItemViewHolder(this);
-                return;
-            }
-            if (mSwipeRefresh.isRefreshing()) {
-                return;
-            }
-            if (mItem == null) {
-                return;
-            }
-
-            if (mItem.getState() == BoxListItem.State.ERROR) {
-                mItem.setState(BoxListItem.State.SUBMITTED);
-                getController().execute(mItem.getRequest());
-                setLoading();
-            }
-
-            if (mListener != null) {
-                if (mListener.handleOnItemClick(mItem.getBoxItem())) {
-                    FragmentActivity activity = getActivity();
-                    if (activity == null) {
-                        return;
-                    }
-
-                    if (mItem.getBoxItem() instanceof BoxFolder) {
-                        // Find fragment container id
-                        ViewGroup fragmentContainer = (ViewGroup) mRootView.getParent();
-                        if (fragmentContainer != null) {
-                            BoxFolder folder = (BoxFolder) mItem.getBoxItem();
-                            FragmentTransaction trans = activity.getSupportFragmentManager().beginTransaction();
-
-                            // All fragments will always navigate into folders
-                            BoxBrowseFolderFragment browseFolderFragment = new BoxBrowseFolderFragment
-                                    .Builder((BoxFolder) folder, mSession).build();
-
-                            trans.replace(fragmentContainer.getId(), browseFolderFragment)
-                                    .addToBackStack(TAG)
-                                    .commit();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    protected class BoxItemAdapter extends RecyclerView.Adapter<BoxItemViewHolder> {
-        protected ArrayList<BoxListItem> mListItems = new ArrayList<BoxListItem>();
-        protected HashMap<String, BoxListItem> mItemsMap = new HashMap<String, BoxListItem>();
-
-
-        @Override
-        public BoxItemViewHolder onCreateViewHolder(ViewGroup viewGroup, int i) {
-            View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.box_browsesdk_list_item, viewGroup, false);
-            return new BoxItemViewHolder(view);
-        }
-
-        @Override
-        public void onBindViewHolder(BoxItemViewHolder boxItemHolder, int i) {
-            BoxListItem item = mListItems.get(i);
-            if (item.getState() == BoxListItem.State.ERROR) {
-                boxItemHolder.setError(item);
-                return;
-            }
-            boxItemHolder.bindItem(item);
-
-            // Fetch thumbnails for media file types
-            if (item.getBoxItem() instanceof BoxFile && ThumbnailManager.isThumbnailAvailable(item.getBoxItem())) {
-                if (item.getRequest() == null) {
-                    item.setRequest(getDownloadThumbnailTask(item.getBoxItem().getId(), mThumbnailManager.getThumbnailForFile(item.getBoxItem().getId())));
-                } else if (item.getResponse() != null) {
-                     BoxException ex = (BoxException) item.getResponse().getException();
-                    if (ex != null && ex.getResponseCode() != HttpURLConnection.HTTP_NOT_FOUND) {
-                        item.setState(BoxListItem.State.CREATED);
-                    }
-                }
-            }
-
-            // Execute a request if it hasn't been done so already
-            if (item.getRequest() != null && item.getState() == BoxListItem.State.CREATED) {
-                item.setState(BoxListItem.State.SUBMITTED);
-                getController().execute(item.getRequest());
-                if (item.getIdentifier().equals(ACTION_FUTURE_TASK)) {
-                    boxItemHolder.setLoading();
-                }
-                return;
-            }
-        }
-
-        @Override
-        public int getItemCount() {
-            return mListItems.size();
-        }
-
-        public BoxListItem get(String id) {
-            return mItemsMap.get(id);
-        }
-
-        public synchronized void removeAll() {
-            mItemsMap.clear();
-            mListItems.clear();
-        }
-
-        public void remove(BoxListItem listItem) {
-            remove(listItem.getIdentifier());
-        }
-
-        public synchronized void remove(String key) {
-            BoxListItem item = mItemsMap.remove(key);
-            if (item != null) {
-                int index = item.getPosition() != null ?
-                        item.getPosition().intValue() :
-                        mListItems.indexOf(item);
-                mListItems.remove(index);
-                this.notifyItemRemoved(index);
-            }
-        }
-
-        public void addAll(BoxIteratorItems items) {
-            for (BoxItem item : items) {
-                if (!mItemsMap.containsKey(item.getId())) {
-                    add(new BoxListItem(item, item.getId()));
-                } else {
-                    // update an existing item if it exists.
-                    mItemsMap.get(item.getId()).setBoxItem(item);
-                }
-            }
-        }
-
-        public synchronized void add(BoxListItem listItem) {
-            if (listItem.getBoxItem() != null) {
-                // If the item should not be visible, skip adding the item
-                if (!isItemVisible(listItem.getBoxItem())) {
-                    return;
-                }
-                //Filter out item if necessary
-                if (mBoxItemFilter != null && !mBoxItemFilter.accept(listItem.getBoxItem())) {
-                    return;
-                }
-                listItem.setIsEnabled(isItemEnabled(listItem.getBoxItem()));
-            }
-            mListItems.add(listItem);
-            listItem.setPosition(mListItems.size() - 1);
-            mItemsMap.put(listItem.getIdentifier(), listItem);
-        }
-
-        public void update(String id) {
-            BoxListItem item = mItemsMap.get(id);
-            if (item != null) {
-                int index = item.getPosition() != null ?
-                        item.getPosition() :
-                        mListItems.indexOf(item);
-                notifyItemChanged(index);
-            }
-        }
-
-        public void update(BoxItem item) {
-            BoxListItem listItem = mItemsMap.get(item.getId());
-            if (listItem != null) {
-                listItem.setBoxItem(item);
-                int index = listItem.getPosition() != null ?
-                        listItem.getPosition() :
-                        mListItems.indexOf(listItem);
-                notifyItemChanged(index);
-            }
-        }
-    }
-
-    private class BoxItemClickListener implements View.OnClickListener {
-
-        protected BoxListItem mBoxListItem;
-
-        void setListItem(BoxListItem listItem) {
-            mBoxListItem = listItem;
-        }
-
-        @Override
-        public void onClick(View v) {
-            getSecondaryActionListener().handleOnItemClick(mBoxListItem.getBoxItem());
-        }
-
-    }
-
 
     /**
      * Builder for constructing an instance of BoxBrowseFolderFragment
